requirements.txt
pyrogram
tgcrypto
py-tgcalls==0.5.5
python-dotenv
youtube_dl==2021.06.06
youtube_search_python==1.4.6
requests
aiohttp
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg
Pillow==8.3.1
future
psutil
wget
tswift
